Move `KVTransferConfig` from `config/__init__.py` to `config/kv_transfer.py`
